requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
aiogram
apscheduler
httpx
requests